configMongoDB/requirements.txt
pymongo
pyyaml
pyaes
